separation of concerns before adding more components
